Updated the tbb converter and reader programs to handle the arrays as shorts instead of ints to match the spec and save space.
git-svn-id: http://usg.lofar.org/svn/code/trunk/src/DAL@1113 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ implement/dalDataset.h
@@ -46,7 +46,7 @@
 
 /*!
   \class dalDataset
-  
+
   \ingroup DAL
 
   \brief Represents the file containing all sub-structures
@@ -293,7 +293,7 @@ class dalDataset{
 	                starting at the specified start value.
 	  \return A pointer to a structure containing the block of data.
 	 */
-	void read_tbb(string id, int start, int length, int data_out[]);
+	void read_tbb(string id, int start, int length, short data_out[]);
 
 /************************************************************************
  *
